Refactor some modules into `common`, replace all relative imports with absolute imports
