+++ govapp/apps/logs/urls.py
@@ -5,7 +5,7 @@
 from rest_framework import routers
 
 # Local
-from . import views
+from govapp.apps.logs import views
 
 
 # Router
